+++ lib/ptc_runner/lisp/runtime.ex
@@ -354,7 +354,7 @@ defmodule PtcRunner.Lisp.Runtime do
   defdelegate java_util_date(ms), to: Interop
   defdelegate dot_get_time(dt), to: Interop
   defdelegate current_time_millis, to: Interop
-  defdelegate local_date_parse(s), to: Interop
+  defdelegate parse_temporal(s), to: Interop
   defdelegate dot_contains(s, substring), to: Interop
   defdelegate dot_index_of(s, substring), to: Interop
   defdelegate dot_index_of(s, substring, from), to: Interop